Finishing renaming contact data field to contact custom data field
