+++ spec/paths/contacts/contact-custom-data-fields.ts
@@ -1,6 +1,6 @@
 import { PathItemObject, OperationObject, ResponseObject, ResponsesObject } from 'openapi3-ts';
 import { ref as errorResponseRef } from '../../components/responses/error';
-import { ref as contactDataFieldRef } from '../../components/schemas/contacts/contact-data-field';
+import { ref as contactDataFieldRef } from '../../components/schemas/contacts/contact-custom-data-field';
 
 const post: OperationObject = {
   summary: 'Create a new contact custom data field',
